LayoutTests/http/tests/media/media-stream/enumerate-devices-iframe-allow-attribute-expected.txt

CONSOLE MESSAGE: Feature policy 'Camera' check failed for iframe with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Feature policy 'Microphone' check failed for iframe with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Not allowed to call enumerateDevices.
      
PASS: "allow" attribute respected in all iframes
